+++ app/src/main/java/woowacourse/omok/omokdb/OmokDbHelper.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.omokdb
 
 import android.annotation.SuppressLint
 import android.content.ContentValues
